pyproject.toml
[project]
name = "cx-lifecycle"
version = "0.1.0"
description = "Supply Chain Risk Monolith"
authors = [
    {name = "Gemini",email = "gemini@google.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.124.2,<0.125.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "sqlalchemy[asyncio] (>=2.0.45,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "google-genai (>=1.55.0,<2.0.0)",
    "google-cloud-aiplatform (>=1.130.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "greenlet (>=3.3.0,<4.0.0)",
    "dirty-equals (>=0.11,<0.12)",
    "mypy (>=1.19.0,<2.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)"
]
